debbug/requirements_debug.txt
#basico
esptool>=4.6.2
mpremote>=1.22.0

#debug teste
pyserial>=3.5
pytest>=7.0.0
pytest-asyncio>=0.21.0

pylint>=2.15.0
black>=22.0.0
flake8>=5.0.0

requests>=2.28.0
pyyaml>=6.0

#dev ESP32
adafruit-ampy>=1.0.0
rshell>=0.0.30

#monitoramento
psutil>=5.9.0

#rede
netifaces>=0.11.0

#pip install -r requirements_debug.txt